libticonv, libtifiles, libticables, libticalcs, gfm, tilp: fix a number of problems related to unaligned reads/writes and left shifts of count >= width of type; fix more compiler warnings reported by clang. Some problems were flagged by `clang -fsanitize=undefined`, and the same fix was applied to other similar code chunks.
Authors: mainly Lionel Debroux, Benjamin Moody pointed out more code chunks.

git-svn-id: http://svn.tilp.info/repos/tilp@4478

+++ libticables/trunk/src/hex2dbus.c
@@ -204,7 +204,7 @@ restart:
 		mid = buffer[i+0];
 		cid = buffer[i+1];
 		length = buffer[i+2];
-		length |= buffer[i+3] << 8;
+		length |= ((int)(buffer[i+3])) << 8;
 
 		// check for valid packet
 		if(is_a_machine_id(mid) == -1)
